Adding a Hashie masher to convert responses to Hashie::Mash
