+++ lib/shipstation/client/response.rb
@@ -1,0 +1,9 @@
+require 'hashie/mash'
+
+module ShipStation
+  class Client
+    class Response < Hashie::Mash
+      disable_warnings
+    end
+  end
+end
