+++ shipstation.gemspec
@@ -28,4 +28,5 @@ Gem::Specification.new do |spec|
 
 
   spec.add_dependency 'httparty'
+  spec.add_dependency 'hashie'
 end
